requirements.txt
streamlit
langchain
python-dotenv
# Add any other libraries you import (e.g., requests, openai, etc.)